Fix: Stop() isn't invoked when timer's time runs out
When elapsed == total, IsTimeLeft() reports false. This means Stop()
wouldn't signal stopCh since it is protected by a `if IsTimeLeft()`
conditional.

I could eschew IsTimeLeft() and use `if c.elapsed <= c.total`, but that
is duplicating logic, which I want to avoid.

I could also compeletely remove the if statement, but the problem with
that approach is that Stop() would signal stopCh even when timer isn't
running.

So, the best approach I see is to add a `running` field to the Clock.
This also allows me to implement Reset().