Implement Select() for Queue
Planning to add a Next() and Prev() functioin to the Queue. They would
need to select the next and prev item from the Queue, while making sure
that the head of the Queue also moves along with it. Without a Select(),
functionality would be duplicated between Next(), Prev() and the handler
installed via Queue's Table.SetSelectedFunc().