Remove Format field from Queue
Queue has items added to it only at it's initialisation. No queue items
can be added during it's lifetime. This means setting Queue.Format after
NewQueue() has no effect on the text in the duration column. So, a new
format has to be applied to all of the cells all at once with a
function. This eliminates any use case for storing the formatting
function.