Add type ButtonColumn
I need this to render buttons the way I want them to be, which is

> | Button 1 | | Button 2 | ... | Button N |

With uniform width for all of the buttons, and a constant height. A Grid
can't do fixed size without some row/colSpan magicking and a Flex can
restrict only the width, not the height (with FlexColumn). You also have
to calculate the button width upfront while I believe that this should
be handled by the widget.