Implement Started() and Stopped() handlers for Clock
This would allow me to change the label of the play/pause buttons
whenever the state of the Clock changed. Previously, I did this by
running a gorouting that would compare the label and the state of Clock
to determine what label to set for the button, but that was hell on my
CPU. CPU usage climbed to 30% with that solution, so I've decided to
ditch that approach for this one (even though I don't really like
modifying Clock to add yet more handlers).